WIP:SingleStintPlayer_12-Class
Created PersonSingleStintRecord (extension of PersonStint) & PersonSingleStintList. Moved some work from the GUI project to the new classes (particularly checking whether the single stint is already editable & recording it as a bool) & implemented better progress reporting for the initial creation of the list. Removing manually calculated info from the DataGridView and moving it to a dedicated class will make implementing the filters much easier for the next update.